Orientation handling revamp setup:
Swapped all calls of UpdatedComponent->GetUpVector() to a function GetUpOrientation(). Mainly setup, introduced  gravity variable to be able to get directions from it since it's made air ticks trickier as player orientation may not be reliable. The method will return player up vector if gravity is near zero though as a fall back.

The 3 possible return values intended are, PlayerUp, -GravityDir, FloorNormal. Depending on the situation and argument. Right now nothings changed, but plan on going into it with the arguments on a case-by-case basis for all references needing an up vector. Will see what makes sense where.